Rewrote

+++ pinbot.py
@@ -27,12 +27,8 @@ botName = "pinbot" # Your bot's name goes here
 botCreator = "pinhead#4946"
 ver = ['v1.0.0', '06-06-2021'] # Bot version and release date goes here, major update.minor update.bugfix update.
 config = 'config.yml' # Open this file and put your bot's name and token there
-
 guild_ids = [827647182051737651, 851199078574718982] # Server ID goes here
 sayPermIDs= [246291288775852033, 495303865214959618, 474759210056548353, 488414757628411934, 638864530964742184] # People you want to give say perms go here
-botLoggingChannel = discord.utils.get(bot.get_all_channels(), name="pinbot-logs") # Channel to log specific commands and other info in
-botInbox = discord.utils.get(bot.get_all_channels(), name="pin-box") # Channel to put adverts, modmails, and any other features I decide to add
-botCommandChannel = discord.utils.get(bot.get_all_channels(), name="pinbot-commands") # Staff commands channel
 botStartTime = DT.datetime.now()
 
 lout.writeFile(botName + 'Logs.txt', '\n' + botName + 'Initialized Successfully!', True)
@@ -125,7 +121,9 @@ async def help(ctx):
     embed.add_field(name=botPrefix + 'time, serverTime', value='States what time it is on the server that the bot is hosted on', inline=False)
     embed.add_field(name=botPrefix + 'uptime, runTime, runDuration', value=botName + ' reports how long it has run without going offline.', inline=False)
     if role in ctx.author.roles: # If user has the role you defined, show commands in the if statement, highly recommended you put staff only commands here
-        embed.add_field(name=botPrefix + 'mail, sendMail, dm', value='Command used to send modmail to users. Staff only command. Command goes as follows, and yes "" is required: mail 246291288775852033 "Modmail content"', inline=False)
+        embed.add_field(name=botPrefix + 'mail, sendMail, dm', value='Command used to send modmail to users. Staff only command. Command goes as follows, and yes "" is required: mail 246291288775852033 Hello!', inline=False)
+    if ctx.author.id in sayPermIDs:
+        embed.add_field(name=botPrefix + 'say', value='Forces ' + botName + ' to send a message. Command is as follows: p!say Hello There', inline=False)
     embed.add_field(name=botPrefix + 'shutdown, stop, kill', value='Shuts down the bot safely and outputs it to logs.', inline=False)
     embed.set_footer(text=botName + ' ' + botVersion + ' | ' + botVersionReleaseDate + ' | Created by pinhead [' + botCreator + ']') # Your name goes in the place of pinhead
     await author.send(embed=embed)
@@ -138,17 +136,26 @@ async def help(ctx):
 @bot.command(pass_context=True)
 async def say(ctx, *messageContent):
     startTime = DT.datetime.now()
-    if ctx.author.id in sayPermIDs:
-        await ctx.send(" ".join(messageContent[:]))
-        msg = await ctx.fetch_message(ctx.message.id)
-        await msg.delete()
-    else:
+    if messageContent == []:
         embed = discord.Embed(
             color = discord.Color.red()
         )
-        embed.add_field(name='An error occured while executing the command "say". The error is as follows:', value="```diff \n - You don't have say perms!```")
-        await ctx.send(embed=embed)
-    lout.log(config, startTime, 'say')
+        embed.add_field(name='An error occured while executing the command "say". The error is as follows:', value='```diff\n - The message content is empty.\n + Command usage: say Hello!```')
+        lout.log(config, startTime, 'say')
+        return await ctx.send(embed=embed)
+    else:
+        if ctx.author.id in sayPermIDs:
+            await ctx.send(" ".join(messageContent[:]))
+            msg = await ctx.fetch_message(ctx.message.id)
+            await msg.delete()
+            lout.log(config, startTime, 'say')
+        else:
+            embed = discord.Embed(
+                color = discord.Color.red()
+            )
+            embed.add_field(name='An error occured while executing the command "say". The error is as follows:', value="```diff\n - You don't have say perms!```")
+            await ctx.send(embed=embed)
+        lout.log(config, startTime, 'say')
 
 #  __            ___  __   __            
 # /__` |__| |  |  |  |  \ /  \ |  | |\ | 
@@ -176,6 +183,7 @@ async def on_message(ctx):
     startTime = DT.datetime.now()
     sender_id = str(ctx.author.id)
     pinbotLogs = discord.utils.get(bot.get_all_channels(), name="pinbot-logs") # Bot log channel goes here
+    pinBox = discord.utils.get(bot.get_all_channels(), name="pin-box") # Bot inbox channel goes here
     if ctx.author == bot.user:
         return
     if str(ctx.channel.type) == "private":  
@@ -184,41 +192,38 @@ async def on_message(ctx):
             color = discord.Color.green()
         )
         embed.add_field(name='User ID:', value=sender_id)
-        await botInbox.send("**New Modmail!**\n" + "**From:** <@" + sender_id + ">!\n" + "**Message: **" + ctx.content, embed=embed)
+        await pinBox.send("**New Modmail!**\n" + "**From:** <@" + sender_id + ">!\n" + "**Message: **" + ctx.content, embed=embed)
         lout.logModmail(config, startTime, 'modmailRecieved', ctx.content)
 
 #  |\/|  /\  | |    
 #  |  | /~~\ | |___ 
 
 @bot.command(pass_context=True, aliases=['sendMail', 'dm'])
-async def mail(ctx, user: discord.User, messageContent: str = None):
+async def mail(ctx, user: discord.User, *messageContent):
     startTime = DT.datetime.now()
     pinbotLogs = discord.utils.get(bot.get_all_channels(), name="pinbot-logs") # Bot log channel goes here
-    # if user != str:
-    #     embed = discord.Embed(
-    #         color = discord.Color.red()
-    #     )
-    #     embed.add_field(name='An error occured while executing the command "mail". The error is:', value='```diff - You need to give a valid user ID.\n+ Command usage: mail 246291288775852033 "Hello!"```')
-    #     lout.log(config, startTime, 'mail')
-    #     return await ctx.send(embed=embed)
-    if messageContent is None:
+    print(user)
+    if messageContent == ():
         embed = discord.Embed(
             color = discord.Color.red()
         )
-        embed.add_field(name='An error occured while executing the command "mail". The error is as follows:', value='```diff \n - The message content is empty.\n + Command usage: mail 246291288775852033 "Hello!"```')
+        embed.add_field(name='An error occured while executing the command "mail". The error is as follows:', value='```diff\n - The message content is empty.\n + Command usage: mail 246291288775852033 Hello!```')
         lout.log(config, startTime, 'mail')
         return await ctx.send(embed=embed)
     role = discord.utils.get(ctx.guild.roles, name="Staff") # Your server's staff role goes here
     if role in ctx.author.roles:
         if str(ctx.channel) == "pinbot-commands": # Staff commands channel goes here
-            await user.send("**New Modmail from** <@" + str(ctx.author.id) + ">!\n" + messageContent)
-            await pinbotLogs.send("**Modmail sent to** <@" + str(user.id) + ">.\n**Message sent:** " + messageContent)
+            sendToUser = "**New Modmail from** <@" + str(ctx.author.id) + ">!\n"
+            sendToLogs = "**Modmail sent to** <@" + str(user.id) + ">.\n**Message sent:** "
+            print(messageContent)
+            await user.send(sendToUser + " ".join(messageContent[:]))
+            await pinbotLogs.send(sendToLogs + " ".join(messageContent[:]))
             lout.logModmail(config, startTime, 'modmailSent', messageContent)
     else:
         embed = discord.Embed(
             color = discord.Color.red()
         )
-        embed.add_field(name='An error occured while executing the command "mail". The error is as follows:', value="```diff \n - You're not staff!```")
+        embed.add_field(name='An error occured while executing the command "mail". The error is as follows:', value="```diff\n - You're not staff!```")
         lout.log(config, startTime, 'mail')
         return await ctx.send(embed=embed)
